broke my file system...

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: alemarch <alemarch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/17 00:18:10 by alemarch          #+#    #+#             */
-/*   Updated: 2021/11/20 17:08:28 by alemarch         ###   ########.fr       */
+/*   Updated: 2021/11/22 14:17:40 by alemarch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(void *s1, void *s2, size_t n)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
